+++ backend/src/main/java/com/betolyn/features/betting/betslips/dto/BetSlipItemDTO.java
@@ -3,6 +3,8 @@ package com.betolyn.features.betting.betslips.dto;
 import com.betolyn.features.betting.betslips.enums.BetSlipItemStatusEnum;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class BetSlipItemDTO {
 
@@ -13,9 +15,9 @@ public class BetSlipItemDTO {
     private String oddId;
     private String lastOddHistoryId;
 
-    private Double stake;
-    private Double potentialPayout;
-    private Double oddValueAtPlacement;
+    private BigDecimal stake;
+    private BigDecimal potentialPayout;
+    private BigDecimal oddValueAtPlacement;
 
     private String voidReason;
