+++ backend/src/main/java/com/betolyn/features/betting/betslips/placebet/PlaceBetItemParam.java
@@ -2,9 +2,11 @@ package com.betolyn.features.betting.betslips.placebet;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PlaceBetItemParam {
     private String oddId;
-    private Double stake;
-    private Double oddValueAtPlacement;
+    private BigDecimal stake;
+    private BigDecimal oddValueAtPlacement;
 }
