+++ backend/src/main/java/com/betolyn/features/betting/criterion/dto/CreateCriterionOddRequestDTO.java
@@ -5,12 +5,14 @@ import com.betolyn.features.betting.odds.OddStatusEnum;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CreateCriterionOddRequestDTO {
     private String name;
 
     @NotNull
-    private double value;
+    private BigDecimal value;
 
     private OddStatusEnum status;
 }
