+++ backend/src/main/java/com/betolyn/features/betting/criterion/dto/CriterionDTO.java
@@ -1,12 +1,14 @@
 package com.betolyn.features.betting.criterion.dto;
 
+import java.math.BigDecimal;
+import java.util.List;
+
 import com.betolyn.features.betting.criterion.CriterionStatusEnum;
 import com.betolyn.features.betting.odds.dto.OddDTO;
 import com.betolyn.features.matches.MatchDTO;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Data;
 
-import java.util.List;
+import lombok.Data;
 
 @Data
 public class CriterionDTO {
@@ -16,10 +18,10 @@ public class CriterionDTO {
     private boolean allowMultipleWinners;
     private boolean isStandalone;
 
-    private Double totalBetsCount;
-    private Double totalStakesVolume;
-    private Double reservedLiability;
-    private Double maxReservedLiability;
+    private Integer totalBetsCount;
+    private BigDecimal totalStakesVolume;
+    private BigDecimal reservedLiability;
+    private BigDecimal maxReservedLiability;
 
     private MatchDTO match;
     private CriterionStatusEnum status;
