+++ backend/src/main/java/com/betolyn/features/betting/odds/dto/UpdateOddRequestDTO.java
@@ -3,8 +3,10 @@ package com.betolyn.features.betting.odds.dto;
 import com.betolyn.features.betting.odds.OddStatusEnum;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class UpdateOddRequestDTO {
     private OddStatusEnum status;
-    private Double value;
+    private BigDecimal value;
 }
