+++ backend/src/main/java/com/betolyn/features/matches/getmatchmetrics/MatchMetricsDTO.java
@@ -1,5 +1,7 @@
 package com.betolyn.features.matches.getmatchmetrics;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -10,13 +12,13 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class MatchMetricsDTO {
-    private Double totalVolume;
-    private Double reservedLiability;
-    private Double maxReservedLiability;
+    private BigDecimal totalVolume;
+    private BigDecimal reservedLiability;
+    private BigDecimal maxReservedLiability;
     /** Risk level 0–100 from reservedLiability / maxReservedLiability; 0 when max is null or 0. */
     private Double riskLevel;
     private Integer totalCriteriaCount;
-    private Long totalBetCount;
+    private Integer totalBetCount;
     /** Realized P/L (after settlement); null when no settled criteria. */
-    private Double profitAndLosses;
+    private BigDecimal profitAndLosses;
 }
